Renamed the scanner framework file.
