+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/aila-scanner-cordova/aila-to-openpos-barcode-type.ts
@@ -1,5 +1,5 @@
 import { AilaBarcodeType } from './aila-barcode-type.enum';
-import { ScanDataType } from '../image-scanner';
+import { ScanDataType } from '../scanner';
 
 export class AilaBarcodeUtils {
