+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/console-scanner/console-scanner.plugin.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 
-import { Scanner, ScanData, ScanDataType } from '../image-scanner';
+import { Scanner, ScanData, ScanDataType } from '../scanner';
 
 declare global {
     interface Console {
