+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/infinea-scanner-cordova/infinea-to-openpos-barcode-type.ts
@@ -1,4 +1,4 @@
-import { ScanDataType } from '../image-scanner';
+import { ScanDataType } from '../scanner';
 import { InfineaBarcodeType } from './infinea-barcode-type.enum';
 
 export class InfineaBarcodeUtils {
