+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/scandit-scanner-cordova/scandit-to-openpos-barcode-type.ts
@@ -1,4 +1,4 @@
-import { ScanDataType } from '../image-scanner';
+import { ScanDataType } from '../scanner';
 import { Barcode } from './types/barcode';
 
 export class ScanditBarcodeUtils {
