+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/server-scanner/server-scanner.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { SessionService } from '../../../services/session.service';
-import { ScanData, Scanner } from '../image-scanner';
+import { ScanData, Scanner } from '../scanner';
 
 @Injectable({
     providedIn: 'root',
