+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/image-scanners/wedge-scanner/wedge-scanner.plugin.ts
@@ -4,7 +4,7 @@ import { map, filter, bufferToggle, timeout, catchError, windowToggle, tap, merg
 import { SessionService } from '../../../services/session.service';
 import { WEDGE_SCANNER_ACCEPTED_KEYS } from './wedge-scanner-accepted-keys';
 import { DomEventManager } from '../../../services/dom-event-manager.service';
-import { Scanner, ScanData, ScanDataType} from '../image-scanner';
+import { Scanner, ScanData, ScanDataType} from '../scanner';
 
 interface ControlSequence { modifiers: string[]; key: string; }
